Add an option to omit nodes set as top-level
Omitting the nodes means they and their children won't be included in the AABB.

These nodes have a fixed position and will ignore the preview node's positioning during drag-and-drop instantiation, so we want to be able to ignore their bounding box contributions in that situation. However, we still want to be able to draw selection boxes around them, which is why the behavior is optional.